feat: Add lazy loading for images
This commit adds lazy loading functionality to the webpage by modifying the HTML file `8-main.html` and the JavaScript file `8-script.js`. The HTML file includes the necessary attributes and elements to enable lazy loading for images. The JavaScript file contains a jQuery script that makes an AJAX GET request to fetch the data of Star Wars movies from the SWAPI API and dynamically generates `<li>` elements to display the movie titles.