feat: Add jQuery script to fetch and display translated greeting
This commit adds a new HTML file `102-main.html` and a new JavaScript file `102-script.js`. The HTML file includes a header, input fields, and a div element to display the translated greeting. The JavaScript file contains a jQuery script that makes an AJAX GET request to the URL `https://www.fourtonfish.com/hellosalut/?` with the language code entered by the user. The response data is then used to update the content of the div element with the translated greeting.